// src/app/shared/models/sales/productivity.ts
export class Productivity {
    area: string;
    customer: string;
    vicem01: number;
    vicem02: number;
    insee01: number;
    insee02: number;
    sutu01: number;
    sutu02: number;
    diemHuy01: number;
    diemHuy02: number;
    thinhPhat01: number;
    thinhPhat02: number;
    // thinhPhatTotal: number;
    total: number;
    items: any[];
}